Create SplunkMonthlyQueries.py
This is the beginnings of a project that stemmed from Splunk Admins refusing to give the Security Operations Center any access to schedule queries or reports and because they sucked at configuring and building it out correctly.  So we came up with an architecture that would allow us to write the query in a file, place it in a directory and then this script could run an authenticated process (controlled by the almighty developers/admins).  They would schedule Splunk to call this Python script once a month, it then would parse through the script directory and run our monthly queries.  There is still more to be done on it, but it should get you started if you need it.