Update ShodanTrojans.py
This script was designed to be run 'on-demand' to pull data from Shodan using the API where the following web query could be used product:*trojan* via the web interface.  It puts the full data set into ShodanResults_YYYY-MM-DD.txt file and creates two other text files with just the IPs and just the domains.   

It should be noted that I did make a minor modification to some of the shodan package for error statements where I changed syntax from SomeError, e: to SomeError as e: to be compatible with 3.4 formatting.