Create DomainTools_ProcessDomainHistory.py
After using this script you can use the unique_ips.txt file with the ShodanBulkHost search script and get the historical data for all the IPs.  You should make a separate folder for the Shodan output.  After that, you can take it a step further and use the ShodanExtractElements script to walk the master file and pull out specific entries for the meta data and write them to one file.