Added Shodan
Some users will not be able to afford the MaxMind resources so they will need to pull from Shodan instead.  For those who can, then this just gives a secondary source to which they can compare the MaxMind data returned in addition, Shodan offers data elements regarding a host that MaxMind does not and this just providers a richer data set.