get_creds function added
added a function to pull various credentials from a pre-defined path so that they don't have to be placed directly in the script and can be unique to each user running the script.

added code to search for a default file structure for input and config data for the script, and to create it if it does not exist.  Leaving the user to only have to paste the IP addresses into the text file and to place the proper GeoIP2-City.mmd database in the appropriate directory.  Significant comments created for user to read and comprehend as well.