Create cves_exploits.py
This script was written to serve a single purpose, identify whether or not there is an exploit or metasploit module out there for a specific CVE number without having to pull any repositories, search multiple sources, or have any local installations etc.  Too often vendors choose not to patch something and forget about the CVE or orgs choose not to apply patches and then forget the vuln exists.  This is an easy way to go check and see if there is now an exploit for those CVEs you've been ignoring and pretending don't exist in your environment.  It is crude, and can be improved, but it works.